+++ lib/engineyard-serverside/deprecation.rb
@@ -1,7 +1,18 @@
+require 'engineyard-serverside/shell/helpers'
+
 module EY
   module Serverside
     def self.deprecation_warning(msg)
       $stderr.puts "DEPRECATION WARNING: #{msg}"
+    end
+
+    def self.const_missing(const)
+      if const == :LoggedOutput
+        EY::Serverside.deprecation_warning("EY::Serverside::LoggedOutput has been deprecated. Use EY::Serverside::Shell::Helpers instead.")
+        EY::Serverside::Shell::Helpers
+      else
+        super
+      end
     end
   end