+++ spec/basic_deploy_spec.rb
@@ -12,17 +12,17 @@ describe "Deploying an application without Bundler" do
 
     # run a deploy
     config = EY::Serverside::Deploy::Configuration.new({
-        "strategy"      => "IntegrationSpec",
-        "deploy_to"     => @deploy_dir.to_s,
-        "group"         => `id -gn`.strip,
-        "stack"         => 'nginx_passenger',
-        "migrate"       => nil,
-        'app'           => 'foo',
-        'framework_env' => 'staging'
-      })
+      "strategy"      => "IntegrationSpec",
+      "deploy_to"     => @deploy_dir.to_s,
+      "group"         => `id -gn`.strip,
+      "stack"         => 'nginx_passenger',
+      "migrate"       => nil,
+      'app'           => 'foo',
+      'framework_env' => 'staging'
+    })
 
     @binpath = File.expand_path(File.join(File.dirname(__FILE__), '..', 'bin', 'engineyard-serverside'))
-    @deployer = FullTestDeploy.new(config)
+    @deployer = FullTestDeploy.new(config, test_shell)
     @deployer.deploy
   end