+++ spec/spec_helper.rb
@@ -17,37 +17,10 @@ module EY
       @node = nil
     end
 
-    module LoggedOutput
-      def self.enable_actual_info!
-        @use_actual_info = true
-      end
-
-      def self.disable_actual_info!
-        @use_actual_info = false
-      end
-
-      def self.use_actual_info?
-        @use_actual_info
-      end
-
-      alias old_info info
-      def info(*args)
-        if EY::Serverside::LoggedOutput.use_actual_info?
-          old_info(*args)
-        end
-      end
-
-      def logged_system(cmd)
-        output = `#{cmd} 2>&1`
-        successful = ($? == 0)
-        if ENV['VERBOSE']
-          if successful
-            $stdout.puts "#{cmd}\n#{output.strip}".chomp
-          else
-            $stderr.puts "\nCommand `#{cmd}` exited with status #{$?.exitstatus}: '#{output.strip}'"
-          end
-        end
-        successful
+    class Shell
+      def spawn_process(cmd, cmd_stdout, cmd_stderr)
+        cmd_stdout << `#{cmd} 2>&1`
+        $? == 0
       end
     end
 
@@ -64,8 +37,7 @@ FileUtils.rm_rf GITREPO_DIR if File.exists? GITREPO_DIR
 Kernel.system "tar xzf #{GITREPO_DIR}.tar.gz -C #{FIXTURES_DIR}"
 
 Spec::Runner.configure do |config|
-  `which npm 2>&1`
-  $NPM_INSTALLED = ($? == 0)
+  $NPM_INSTALLED = system('which npm 2>&1')
   unless $NPM_INSTALLED
     $stderr.puts "npm not found; skipping Node.js specs."
   end
@@ -76,10 +48,46 @@ Spec::Runner.configure do |config|
     EY::Serverside.dna_json = {}.to_json
   end
 
+  class VerboseStringIO < StringIO
+    def <<(str)
+      if ENV['VERBOSE'] || ENV['DEBUG']
+        $stderr << str
+      end
+      super
+    end
+  end
+
+  def stdout
+    @stdout ||= VerboseStringIO.new
+  end
+
+  def stderr
+    @stderr ||= VerboseStringIO.new
+  end
+
+  def read_stdout
+    stdout.rewind
+    stdout.read
+  end
+
+  def read_stderr
+    stderr.rewind
+    stderr.read
+  end
+
+  def read_output
+    read_stdout + "\n" + read_stderr
+  end
+
+  def test_shell
+    log_path =  Pathname.new(Dir.tmpdir).join("serverside-deploy-#{Time.now.to_i}-#{$$}.log")
+    EY::Serverside::Shell.new(:verbose => true, :log_path => log_path, :stdout => stdout, :stderr => stderr)
+  end
+
   def deploy_test_application(assets_enabled = true, &block)
     $DISABLE_GEMFILE = false
     $DISABLE_LOCKFILE = false
-    @deploy_dir = File.join(Dir.tmpdir, "serverside-deploy-#{Time.now.to_i}-#{$$}")
+    @deploy_dir = Pathname.new(Dir.tmpdir).join("serverside-deploy-#{Time.now.to_i}-#{$$}")
 
     # set up EY::Serverside::Server like we're on a solo
     EY::Serverside::Server.reset
@@ -88,7 +96,7 @@ Spec::Runner.configure do |config|
     # run a deploy
     @config = EY::Serverside::Deploy::Configuration.new({
       "strategy"      => "IntegrationSpec",
-      "deploy_to"     => @deploy_dir,
+      "deploy_to"     => @deploy_dir.to_s,
       "group"         => `id -gn`.strip,
       "stack"         => 'nginx_passenger',
       "migrate"       => "ruby -e 'puts ENV[\"PATH\"]' > #{@deploy_dir}/path-when-migrating",
@@ -97,16 +105,16 @@ Spec::Runner.configure do |config|
     })
 
     # pretend there is a shared bundled_gems directory
-    FileUtils.mkdir_p(File.join(@deploy_dir, 'shared', 'bundled_gems'))
+    @deploy_dir.join('shared', 'bundled_gems').mkpath
     %w(RUBY_VERSION SYSTEM_VERSION).each do |name|
-      File.open(File.join(@deploy_dir, 'shared', 'bundled_gems', name), "w") { |f| f.write("old\n") }
+      @deploy_dir.join('shared', 'bundled_gems', name).open("w") { |f| f.write("old\n") }
     end
 
     # Set up the application directory to have the requested asset options.
     prepare_rails31_app(assets_enabled)
 
     @binpath = File.expand_path(File.join(File.dirname(__FILE__), '..', 'bin', 'engineyard-serverside'))
-    @deployer = FullTestDeploy.new(@config)
+    @deployer = FullTestDeploy.new(@config, test_shell)
     @deployer.deploy(&block)
   end